requirements-dev.txt
allennlp==2.9.3
allennlp-models==2.9.3
attrs==21.4.0
black==22.3.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cymem==2.0.6
dnspython==2.2.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
neo4j==4.4.1
nltk==3.7
numpy==1.22.1
packaging==21.3
pandas==1.3.5
pathy==0.6.1
pluggy==1.0.0
preshed==3.0.6
py==1.11.0
py2neo==2021.2.3
pymongo==4.0.1
pydantic==1.8.2
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
seqeval==1.2.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
Scrapy==2.5.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
transformers==4.16.2
typer==0.4.1
typing-extensions==4.0.1
urllib3==1.26.8
wasabi==0.9.0
Werkzeug==2.0.2
wheel==0.37.1
lemminflect==0.2.2
neotime==1.7.4
-f https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0
